+++ changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Turn off the NetBSD 11 heartbeat watchdog, which panics an emulated guest that
+  is merely slow
 
 ## [0.7.0] - 2026-08-09
 ### Added